Add support for Surface Centroid Differentiation
This commit adds full support for the differentiation of
the surface centroid position with regard to the positions of
all the vertices that make it.

This is done through a custom class that handles the interface with
those the MATLAB generated c code and only require vectors of pointers
to coordinates.