Triangulate features for volume calculations
This commit marks all the features required for triangulation are
implemented. Triangulations can also be displayed directly as felineseg.
This is refelected in the change of layout. Some efficiency gains are
certainly possible on the 'intermediate' type of triangulation which
is not maintained just re-generated every step.

Now need to track which vertices are internal to the snake object to
then calculate the volume.