Define relative left/right volumes of surfaces
This commit implements method OrientSurfaceVolume which allows
surfs.voluind to be ordered consistently. This is done to achieve
the correct triangulation influences when calculating areas and
volumes and associated derivatives.

The relative face orientation is implemented using flooding of
the surfaces through edge/volume pairs. Flooded cells are expected
to be contra-rotating to obtain the same direction.