Define Absolute left/right volume adjacency
This implements absolute adjacency. This means that the normal vector to
a surface as defined by the cross product will point for a surface
.voluind=[c0 c1] from c0 to c1.

This is done  using geometric properties floating point mathematics.
For a mesh this is done through the centre of the volume. For snakes using
the "outwards pointing property".