Add tracking of influences of triangulations
The triangulations now store information regarding their influence
on the cells (i.e wether calculation will lead to a positive or
negative influence).

This is done using by matching the order of the base surfaces and
storing the ordering of voluind of the parent surf when the
triangulation is carried out.