Add tracking of grid vertices in the snake
This commit adds a property to snake which handles the tracking
of internal vertices. It is being edited in snakeengine at spawning
and snake concatenation. It should correctly keep track of the inner
vs outer vertex.

TEST still needs to be written in the form of a visual check of
inner vertices for tecplot.