Detect surfaces which are on the outer volume border
This allows the mesh class to return the list of surfaces which exist
at the border of two parent mesh volumes and comparing the volumes
of these cells.

This is done by looking for surfaces with multiple parents and
comparing the volumes in those parents looking for an empty one
next to a full one.