Fix some connectivity error
This error is due to two snaxels being connected by the same edge
on both side of the snaxel. This leads to the snaxel being replaced
twice and dangling edges.

Partial fix by removing the second occurence and letting it appear
on the next step. However this leaves some dangling vertices which
do not get deleted.