Fix connectivity issues
Fix the evolution of connectivity to allow random velocity field to be
sustained for 200 steps. Remains a volume orientation error. Likely due
to the size of the block.

These errors are fixed by only collapsing vertices connected along an edge
one by one. This avoids triplets being generated and removes the need for
the NaboveN condition.