Reinstall RSVS process and agressive spawn.
After connectivity issues, reactivate the RSVS process. This is supported
by an agressive spawning process which lets snaxels spawn this is to avoid
premature convergence.

if two of them are converging on a grid vertex the closest one is moved
and labelled for spawning. Resinstallation is done by adding the SQP
calc back and some small modifications to get it to work. Removed
Nan support and fixed the Lagrangian process.

A bug is identified in the interface, triangulation. The corners