Mid debug of border issues
A lot has been learnt the issue is due to large gradients computed
by the derivatives. These are part of the calculation but in all
apparence should cancel out. As they currently do not the gradient
of pixels lying at the boundary between cells causes significant
instability.

Try a finite differencing of the process to see the actual value.