Remove "json.hpp" include in the "parameters.hpp"
The goal is to streamline some of the includes and avoid the spreading
of the JSON for Modern C++ through the code, causing an increase in
compile and link overheads.

This commit removes that include and leaves it as
an implementation feature used by "parameters.cpp" to read and write
the parameter files.