Fix segfault in tetgen deallocation
In it's deallocation tetgen does not guarantee a clean exit.
Upon destruction of tetgenio object a segfault is created by
facets holelist. This can happen for any element with nested
allocation as they do not hold a destructor. This is handled
in the RSVS framework by the tetgen::io_safe class which now
sets NULL pointers.