Initial implementation of Voronoi to mesh class
This commit adds support for the translation of tetgen's voronoi diagrams to
the native RSVS mesh format. The main difficulty in this process is that
Voronoi diagrams are unbounded and not closed. This initial commit
closes it by stepping 0.2 along "rays" (infinite edges). This will likely
be modified in a future release.

This is untested at the moment. Need to add support for a bounding box.