Extend voronois to mesh bound
This commit extend the voronoi mesh to the external bounding box of the
the tetrahedralised points. This is to create a nice voronoi mesh.
During integration consider developing meshes that where the box points
are alway included as they would be for a surrogate mode.

This is done as this is the furthest the box can be extended while still
not worrying about self intersection behaviour. Also removes the potential
0 length edges that can arise.