Correct removal of singular connectors
The previous version of the code would leave duplicate edges and surfaces in the
mesh. Leading to crashes. These are now correctly removed in
mesh::RemoveSingularConnectors.

This is done by comparing connectivity informations to decide if deletion is
warranted for edges and surfaces. Required some additional implementation in
the surf, edge and mesh containers to handle comparisons and removals.