Handle reallocation of surfaces (major bug)
This commit fixes a pernicious but rare bug when surf::OrderEdges
attempts to create a new surface by splitting the surface. If the
container is at capacity this will trigger reallocation of the
mesh::surfs container and the invalidation of the *this pointer
in the call to surf::OrderEdges. This causes seg faults which are
bad, or unsafe read/writes (which are worse).