Bug in mesh::AddBoundary detected
This commit adds an error message and a partial solution
in mesh::AddBoundary. This bug is when a surface is cut twice
by two cutting conditions and need to be split in more than
two parts (a square where two corners get clipped).

This case will be fixed in the next commit.