Refactor surf::OrderEdges
Surf orderdeges was one of the most bloated functions in the code
base. It is now streamlined to be more representative of functionalities.

The first functionality is provided by function OrderEdgeList. This
provides ordering of edge lists according to the vertices.

The second functionality is to Split the surface if the edge list is
found to be more than a closed loop. This is captured by a new method
surf::SplitSurface.

This refactor should allow use of the two subfunctions in other
areas of the code and improved the clarity of the control flow
of the function.