Handle flexible exports from parameters and json
This commit adds support for a flexible parameter called exportconfig.
This simply expects lists of strings behaving somewhat like a map.

To support SU2 a json translation is created with allows partial
parameter definition into the tetgen::apiparam. This required, to
ensure logical import patterns to modify the location of flatupdate
and justified the creation of a specific rsvsjson.hpp/cpp interface.