Control usage of surface centroid derivative
Add an option in the calculation object RSVScalc to activate or
deactivate the use of Surface Centroid dPos and HPos. While these seem
to allow much improved convergence they also cause some trouble with
osciallations.

Can be the basis of an algorithm for choosing the correct gradients to
enhance convergence.