Fix Readiness test after PopulateIndices
Fix test failure introduced in ad4bd6cfd07e70a0992d4b98857d0d70096e4a23
The test now failed because checkready and isready did not always match
after populating indices, which is actually fine.

Indeed while the arraystruct is Hashed and Maxindex is set, the elements
may not be ready, (and it may not be possible to test) as only the index
operations have been performed.