Change the test framework to create more subgroups
More subgroups of tests allow a better explanation of what is running/
failing. Includes a group of shorter tests which test all features but
take a reasonable amount of time.

This is to ensure that tests are run more often and clearer where the
failure is appearing.

Also add expected time option to the tests, allowing the user to know
if a test is proceeding as expected or something is hanging.